Refactor LicenseDependencyIssue and VulnerabilityIssue interfaces
Previously the interfaces had attributes that were defined as
optional to allow for the function in fossa.ts to return either.
However it was hacky, but the change to not allow optional
attributes required a whole host of other changes since the
fetchCodeAnalysis was depending on attributes to be optional
in order to be able to return either a Vulnerability or
LicenseDependency. It was either leave fetchCodeAnalysis as one
function but all the rest of the code would have to define the
types to be Vuln | LicDep or just split out fetchCodeAnalysis
into two functions so that there isn't this OR condition. Ended
up choosing to split out the function even though most of
the code is now repeated.