Remove no active file condition in useEffect
The condition would set loading to false and it was causing
white screen to happen if there are no active files. Removing
it will allow the skeleton loader to load until currentRepoId
is set and that will trigger the useEffect hook to run. In
useEffect fetchData will get called and once all the data is
fetched, it will set loading to false.